# Documentazione/OpenAPI.yaml
openapi: "3.0.2"
info:
  title: "Aulery"
  description: "Specifica progetto Aulery"
  version: "1.0"
  termsOfService: "https://aulery.it/terms"
  contact:
    name: "Aulery API by Ansima"
    url: "https://aulery.it/info"
    email: "contact@aulery.it"
  license:
    name: "CC Aulery1.0(CC AU 1.0)"
    url: "https://aulery.it"
...
servers:
  - url: "localhost/api/"
    description: Server LocalHost
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  responses:
    GeneralError:
      description: Errore 500, Errore interno
      content:
        text/plain:
          schema:
            type: string
            example: "Errore 500, Errore interno"
    UnauthorizedError:
      description: Bearer token mancante o non valido
...
paths:
  /rest/auth/login:
    post:
      summary: Login amministratore
      description: Operazione che permette di effettuare il login inserendo con username e password. 
      operationId: doLogin
      tags: 
        - Autenticazione      
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        200:
          description: Login effettuato con successo
          content:
            text/plain:
              schema:
                type: string
          headers:
            Authorization:
              schema:
                type: string
                pattern: 'Bearer [a-z0-9A-Z-]+'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/auth/logout:
    delete:
      operationId: doLogout
      tags: 
        - Autenticazione      
      security:
        - bearerAuth: []
      responses:
        200:
          description: Logout effettuato con successo
        401:
          $ref: '#/components/responses/UnauthorizedError'          
        500:
          $ref: '#/components/responses/GeneralError'
...